+++ app/layout.tsx
@@ -30,14 +30,11 @@ const RootLayout = ({ children }: RootLayoutProps) => (
         enableSystem
         disableTransitionOnChange
       >
-        <Navigation />
-        <div
-          className={cn(
-            'mx-auto grid max-w-2xl gap-12 border-dotted pt-24 pb-16',
-            'sm:border-x sm:px-8 sm:pt-32'
-          )}
-        >
-          <ViewTransition>{children}</ViewTransition>
+        <div className="mx-auto grid max-w-2xl gap-12 border-dotted sm:border-x">
+          <Navigation />
+          <div className="grid gap-12 pb-16 sm:px-8">
+            <ViewTransition>{children}</ViewTransition>
+          </div>
         </div>
         <Toaster />
         <ThemeSwitcher />
